Moved DeleteSelected into Display
