+++ frontend/src/js/tabs/content/Display.tsx
@@ -4,6 +4,7 @@ import { GridSelectionModelChangeParams } from '@material-ui/data-grid';
 
 //Importing from other files in the project
 import { ContentTable } from 'solarspell-react-lib';
+import DeleteSelected from './DeleteSelected';
 import ActionPanel from './ActionPanel';
 import ContentForm from './ContentForm';
 import { Content, Metadata, MetadataType } from 'js/types';
@@ -12,7 +13,7 @@ type DisplayActionProps = {
   onEdit: (item: Content, vals: Partial<Content>) => void
   onDelete: (item: Content) => void
   onView: (item: Content) => void
-  onSelectChange: (selected: Content[]) => void
+  onSelectedDelete: (content: Content[]) => void
 }
 
 type DisplayProps = {
@@ -34,6 +35,16 @@ function Display({
   actions,
 }: DisplayProps): React.ReactElement {
   const [editedContent, setEditedContent] = React.useState<Content | undefined>();
+  const [selected, setSelected] = React.useState<Content[]>([]);
+
+  // Ensures deleted content is cleaned from state
+  React.useEffect(
+    () => {
+      const ids = content.map(c => c.id);
+      setSelected(s => s.filter(c => ids.includes(c.id)));
+    },
+    [content],
+  );
 
   const onEdit_ = React.useCallback(
     (item: Content) => setEditedContent(item),
@@ -54,14 +65,18 @@ function Display({
     (
       content: Content[],
       rows: GridSelectionModelChangeParams,
-    ) => actions.onSelectChange(
+    ) => setSelected(
       content.filter(c => rows.selectionModel.includes(c.id))
     ),
-    [actions.onSelectChange],
+    [setSelected],
   );
 
   return (
     <>
+      <DeleteSelected
+        selected={selected}
+        onDelete={actions.onSelectedDelete}
+      />
       <ContentForm
         metadata={metadata}
         metadataTypes={metadataTypes}
