+++ frontend/src/js/tabs/content/Modal.tsx
@@ -4,7 +4,6 @@ import Box from '@material-ui/core/Box';
 
 //Importing from other files in the project
 import Add from './Add';
-import DeleteSelected from './DeleteSelected';
 import SearchBar from './SearchBar';
 import Display, { DisplayActionProps } from './Display';
 import { Content, Metadata, MetadataType, Query } from 'js/types';
@@ -14,9 +13,7 @@ type ModalProps = {
     metadataTypes: MetadataType[]
     content: Content[]
     actions: {
-        Display: {
-            onSelectedDelete: (content: Content[]) => void
-        } & Omit<DisplayActionProps, 'onSelectChange'>
+        Display: DisplayActionProps
         Toolbar: {
             onAdd: (content: Content) => void
         }
@@ -32,17 +29,6 @@ function Modal({
     content,
     actions,
 }: ModalProps): React.ReactElement {
-    const [selected, setSelected] = React.useState<Content[]>([]);
-
-    // Ensures deleted content is cleaned from state
-    React.useEffect(
-        () => {
-            const ids = content.map(c => c.id);
-            setSelected(s => s.filter(c => ids.includes(c.id)));
-        },
-        [content],
-    );
-
     return (
         <Box p={2}>
             <Add
@@ -55,18 +41,11 @@ function Modal({
                 metadataTypes={metadataTypes}
                 onQueryChange={actions.Search.onQueryChange}
             />
-            <DeleteSelected
-                selected={selected}
-                onDelete={actions.Display.onSelectedDelete}
-            />
             <Display
                 metadata={metadata}
                 metadataTypes={metadataTypes}
                 content={content}
-                actions={{
-                    ...actions.Display,
-                    onSelectChange: setSelected,
-                }}
+                actions={actions.Display}
             />
         </Box>
     );
